genefinding: rename 'run_prodigal' module to 'prodigal'
