+++ antismash/support/genefinding/prodigal.py
